requirements.txt
tensorflow-gpu
opencv-python
filterpy
numba
numpy
matplotlib
scipy